Adopt site interface for build scripts
Context:

Recently, we've added a `site` interface to replace the old
Makefiles to generate min site , so now we don't necessarily need
to copy over these files in each repo and we can also start using
it directly in the actions.

This commit removes the Makefiles and it also updates the build
scripts for those sample and template repos.

Further cleanup:

There are two mainly: first of all we will need to remove those
file references from the `cims.yml` file, so it doesn't try to
copy these over and we also need to look into the CI packages
like `yq`, if we still need those for anything else.